relocate and abstract similar charges modal
