+++ packages/client/src/components/common/modals/index.ts
@@ -17,4 +17,5 @@ export * from './insert-tax-category.js';
 export * from './match-document-modal.js';
 export * from './modal.js';
 export * from './pull-documents-modal.js';
+export * from './similar-charges/index.js';
 export * from './upload-documents-modal.js';
