+++ packages/client/src/components/common/modals/similar-charges/index.ts
@@ -1,0 +1,1 @@
+export * from './similar-charges-by-id-modal.js';
